Moved image utils to 'image.rs'
